+++ src/Streamly/Internal/Data/Stream/Zip.hs
@@ -21,12 +21,6 @@ module Streamly.Internal.Data.Stream.Zip
     , zipWithK
     , zipWithMK
 
-    , ZipAsyncM(..)
-    , ZipAsync
-    , consMZipAsync
-    , zipAsyncWithK
-    , zipAsyncWithMK
-
     -- * Deprecated
     , ZipStream
     )
@@ -49,7 +43,6 @@ import Text.Read
        ( Lexeme(Ident), lexP, parens, prec, readPrec, readListPrec
        , readListPrecDefault)
 import Streamly.Internal.BaseCompat ((#.), errorWithoutStackTrace, oneShot)
-import Streamly.Internal.Control.Concurrent (MonadAsync)
 import Streamly.Internal.Data.Maybe.Strict (Maybe'(..), toMaybe)
 import Streamly.Internal.Data.Stream.Serial (SerialT(..))
 import Streamly.Internal.Data.Stream.StreamK.Type (Stream)
@@ -57,14 +50,10 @@ import Streamly.Internal.Data.Stream.StreamK.Type (Stream)
 import qualified Streamly.Internal.Data.Stream.Prelude as P
     (cmpBy, eqBy, foldl', foldr, fromList, toList)
 import qualified Streamly.Internal.Data.Stream.StreamK.Type as K
-import qualified Streamly.Internal.Data.Stream.StreamK as K
 import qualified Streamly.Internal.Data.Stream.StreamD as D
 import qualified Streamly.Internal.Data.Stream.Serial as Serial
-import qualified Streamly.Internal.Data.Stream.SVar.Eliminate as SVar
-import qualified Streamly.Internal.Data.Stream.SVar.Generate as SVar
 
 import Prelude hiding (map, repeat, zipWith, errorWithoutStackTrace)
-import Streamly.Internal.Data.SVar
 
 #include "Instances.hs"
 
@@ -88,41 +77,6 @@ zipWithMK f m1 m2 =
 zipWithK :: Monad m
     => (a -> b -> c) -> Stream m a -> Stream m b -> Stream m c
 zipWithK f = zipWithMK (\a b -> return (f a b))
-
-------------------------------------------------------------------------------
--- Parallel Zipping
-------------------------------------------------------------------------------
-
--- | Like 'zipAsyncWith' but with a monadic zipping function.
---
--- @since 0.4.0
-{-# INLINE zipAsyncWithMK #-}
-zipAsyncWithMK :: MonadAsync m
-    => (a -> b -> m c) -> Stream m a -> Stream m b -> Stream m c
-zipAsyncWithMK f m1 m2 = K.mkStream $ \st yld sng stp -> do
-    sv <- newParallelVar StopNone (adaptState st)
-    SVar.toSVarParallel (adaptState st) sv $ D.fromStreamK m2
-    K.foldStream st yld sng stp $ K.zipWithM f m1 (getSerialT (SVar.fromSVar sv))
-
--- XXX Should we rename this to zipParWith or zipParallelWith? This can happen
--- along with the change of behvaior to end the stream concurrently.
---
--- | Like 'zipWith' but zips concurrently i.e. both the streams being zipped
--- are evaluated concurrently using the 'ParallelT' concurrent evaluation
--- style. The maximum number of elements of each stream evaluated in advance
--- can be controlled by 'maxBuffer'.
---
--- The stream ends if stream @a@ or stream @b@ ends. However, if stream @b@
--- ends while we are still evaluating stream @a@ and waiting for a result then
--- stream will not end until after the evaluation of stream @a@ finishes. This
--- behavior can potentially be changed in future to end the stream immediately
--- as soon as any of the stream end is detected.
---
--- @since 0.1.0
-{-# INLINE zipAsyncWithK #-}
-zipAsyncWithK :: MonadAsync m
-    => (a -> b -> c) -> Stream m a -> Stream m b -> Stream m c
-zipAsyncWithK f = zipAsyncWithMK (\a b -> return (f a b))
 
 ------------------------------------------------------------------------------
 -- Serially Zipping Streams
@@ -179,48 +133,3 @@ instance Monad m => Applicative (ZipSerialM m) where
 
 FOLDABLE_INSTANCE(ZipSerialM)
 TRAVERSABLE_INSTANCE(ZipSerialM)
-
-------------------------------------------------------------------------------
--- Parallely Zipping Streams
-------------------------------------------------------------------------------
---
--- | For 'ZipAsyncM' streams:
---
--- @
--- (<>) = 'Streamly.Prelude.serial'
--- (<*>) = 'Streamly.Prelude.serial.zipAsyncWith' id
--- @
---
--- Applicative evaluates the streams being zipped concurrently, the following
--- would take half the time that it would take in serial zipping:
---
--- >>> s = Stream.fromFoldableM $ Prelude.map delay [1, 1, 1]
--- >>> Stream.toList $ Stream.fromZipAsync $ (,) <$> s <*> s
--- ...
--- [(1,1),(1,1),(1,1)]
---
--- /Since: 0.2.0 ("Streamly")/
---
--- @since 0.8.0
-newtype ZipAsyncM m a = ZipAsyncM {getZipAsyncM :: Stream m a}
-        deriving (Semigroup, Monoid)
-
--- | An IO stream whose applicative instance zips streams wAsyncly.
---
--- /Since: 0.2.0 ("Streamly")/
---
--- @since 0.8.0
-type ZipAsync = ZipAsyncM IO
-
-consMZipAsync :: Monad m => m a -> ZipAsyncM m a -> ZipAsyncM m a
-consMZipAsync m (ZipAsyncM r) = ZipAsyncM $ K.consM m r
-
-instance Monad m => Functor (ZipAsyncM m) where
-    {-# INLINE fmap #-}
-    fmap f (ZipAsyncM m) = ZipAsyncM $ getSerialT $ fmap f (SerialT m)
-
-instance MonadAsync m => Applicative (ZipAsyncM m) where
-    pure = ZipAsyncM . getSerialT . Serial.repeat
-
-    {-# INLINE (<*>) #-}
-    ZipAsyncM m1 <*> ZipAsyncM m2 = ZipAsyncM $ zipAsyncWithK id m1 m2
